+++ assets/FullCalendarAssets.php
@@ -12,33 +12,11 @@ use yii\web\AssetBundle;
 
 class FullCalendarAssets extends AssetBundle
 {
-    public $publishOptions = [
-        'forceCopy' => false
-    ];
+    public $publishOptions = ['forceCopy' => false];
     
-    public $sourcePath = '@calendar/node_modules/@fullcalendar';
+    public $sourcePath = '@calendar/resources';
 
-    public $css = [
-        'core/main.min.css',
-        'daygrid/main.min.css',
-        'timegrid/main.min.css',
-        'list/main.min.css',
-    ];
+    public $css = ['css/fullcalendar.bundle.min.css'];
 
-    public $js = [
-        'core/main.min.js',
-        'core/locales-all.min.js',
-        'daygrid/main.min.js',
-        'timegrid/main.min.js',
-        'list/main.min.js',
-        'interaction/main.min.js',
-        'moment/main.min.js',
-        'moment-timezone/main.min.js',
-    ];
-
-    public $depends = [
-        MomentAsset::class,
-        MomentTimezoneAsset::class,
-    ];
-
+    public $js = ['js/fullcalendar.bundle.min.js'];
 }